+++ pkg/utils/utils.go
@@ -1,5 +1,7 @@
 package utils
 
+import "reflect"
+
 func Map[T any, U any](slice []T, f func(T) U) []U {
 	result := make([]U, len(slice))
 	for i, item := range slice {
@@ -16,4 +18,8 @@ func Reduce[T any, U any](slice []T, init U, f func(U, T) U) U {
 	}
 
 	return acc
+}
+
+func SizeOf[T any]() int {
+	return int(reflect.TypeFor[T]().Size())
 }
